Update RESS.md
not the best article, -what is Future Friendly? Is it a thing? I looked it up and it sounded very woo woo

re 22 paragraph to 25- i thought the whole point of RD is that it enables you to create images that resize over different devices, he seems to be saying that this is not the case? 

also, i found some of the writing on wikipedia so i had to rewrite it.

I'd be inclined to leave this whole section out

Luke Wroblewski article was also mentioned on wikipedia and is from 2011, really old for a modern tech site. I think you should add something in your article guidelines about using up recent research unless writers are doing a retrospective or historical piece. 

article lacks logic and flow!